use `String#shellsplit` for the arg to `Kernel#exec`
if a command is executed as `Kernel.exec("ruby my-app.rb with args")
it will be ran in a `sh -c "…"`, and we will get the PID of that one,
which results in orphaned processes. `String#shellsplit` results in:
`Kernel.exec("ruby", "my-app.rb", "with", "args")` which does the
right thing.